[REF] pms_w_s: reorganise unit tests
