+++ pms_website_sale/tests/__init__.py
@@ -1,2 +1,3 @@
 from . import test_booking_engine
+from . import test_booking_engine_parser
 from . import test_routes
