+++ pms_website_sale/tests/test_booking_engine.py
@@ -1,3 +1,6 @@
+# Copyright 2023 Coop IT Easy SC
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
+
 from odoo.tests.common import SavepointCase
